Move date formatting into utils and destructure everything in Card
