+++ src/components/Card/Card.tsx
@@ -3,50 +3,31 @@ import { Post } from '../App/App';
 import { NavLink } from 'react-router-dom';
 import 'dayjs';
 import './Card.css';
-import dayjs from 'dayjs';
+import { formatDateCreated } from '../../utils/formatDateCreated';
 
 interface CardProps {
   post: Post;
 }
 
 export const Card: React.FC<CardProps> = ({ post }) => {
-  const formatDateCreated = () => {
-    let now = dayjs();
-    let createdAt = dayjs(post.createdAt);
-    let diff = now.diff(createdAt, 'hours');
-
-    if (diff >= 24) {
-      let time = dayjs(post.createdAt).format('h:mm a');
-      let date = dayjs(post.createdAt).format('MMMM D, YYYY');
-      return `at ${time} on ${date}`;
-    } else if (diff === 0) {
-      return 'just now';
-    } else if (diff === 1) {
-      return `${diff} hr ago`;
-    } else {
-      return `${diff} hours ago`;
-    }
-  };
-
+  const { id, title, user, image, description } = post;
   return (
-    <NavLink to={`/posts/${post.id}`} className='nav-link'>
+    <NavLink to={`/posts/${id}`} className='nav-link'>
       <button className='card'>
-        <h2>{post.title}</h2>
+        <h2>{title}</h2>
         <div className='user-info'>
           <img
             className='avatar'
-            src={post.user.avatar}
-            alt={`${post.user.username}'s avatar`}
+            src={user.avatar}
+            alt={`${user.username}'s avatar`}
           />
           <p>
-            Posted by {post.user.username} {formatDateCreated()}
+            Posted by {user.username} {formatDateCreated(post)}
           </p>
         </div>
         <div className='preview-container'>
-          {!!post.image && (
-            <img className='post-img' src={post.image} alt={`${post.title}`} />
-          )}
-          {!post.image && <p className='post-desc'>{post.description}</p>}
+          {!!image && <img className='post-img' src={image} alt={`${title}`} />}
+          {!image && <p className='post-desc'>{description}</p>}
         </div>
       </button>
     </NavLink>
