+++ src/components/PostDetails/PostDetails.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { GET_POST_DETAILS } from '../../index';
 import { useQuery } from '@apollo/client';
 import { User } from '../App/App';
-import dayjs from 'dayjs';
+import { formatDateCreated } from '../../utils/formatDateCreated';
 interface PostDetailsProps {
   id: string;
 }
@@ -38,6 +38,16 @@ export const PostDetails: React.FC<PostDetailsProps> = ({ id }) => {
       {details && !loading && !error && (
         <>
           <h2>Title: {details.title}</h2>
+          <div className='user-info'>
+            <img
+              className='avatar'
+              src={details.user.avatar}
+              alt={`${details.user.username}'s avatar`}
+            />
+            <p>
+              Posted by {details.user.username} {formatDateCreated(details)}
+            </p>
+          </div>
           <img
             src={details.image}
             alt={`User upload to go with post titled: ${details.title}`}
